Mar 17, 2026, 11:37 PM

+++ ques2class.cpp
@@ -9,7 +9,7 @@ private:
 string brand ;
 double price;
 public:
-Car(string b,int p){
+Car(string b,double p){
 brand=b;
 price=p;
 }
